Adds features to #72
Addded ng-if to index.html to hide topbar/sidebar when
navigating to /wall/display
Added variable root.fullscreenDisabled to $rootScope
Created a new controller for wall display called wallDisplay
CSS fixes
Added animation
