+++ app/index.html
@@ -23,12 +23,12 @@
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.4/js/bootstrap.min.js"></script>
 </head>
 
-<body class="skin-blue sidebar-mini">
-    <div ng-include src="'topnav.html'"></div>
-    <div ng-include src="'sidebar.html'"></div>
-    <div class="wrapper" ui-view></div>
+<body class="skin-blue sidebar-mini fixed">
+    <div ng-include ng-if="root.fullscreenDisabled" src="'topnav.html'"></div>
+    <div ng-include ng-if="root.fullscreenDisabled" src="'sidebar.html'"></div>
+    <div  ui-view></div>
     <div ng-include src="'pswp.html'"></div>
-    <div ng-include src="'controlsidebar.html'"></div>
+    <div ng-include ng-if="root.fullscreenDisabled" src="'controlsidebar.html'"></div>
     <script src="http://cdnjs.cloudflare.com/ajax/libs/masonry/3.1.2/masonry.pkgd.min.js"></script>
     <script src="http://cdnjs.cloudflare.com/ajax/libs/jquery.imagesloaded/3.1.8/imagesloaded.pkgd.min.js"></script>
     <script src="js/main.js"></script>
