+++ app/styles/main.scss
@@ -600,6 +600,7 @@
 	}
 	.avatar {
 		height: 130px;
+		max-width: 130px;
 		float: left;
 		border-radius: 6px;
 	}
